requirements.txt
flask
pytest
flask-sqlalchemy
psycopg2-binary
python-dotenv
alembic